requirements.txt
chainlit
httpx
langchain
langchain-google-community
langchain-core
langchain-openai
google-api-python-client
google-auth-oauthlib
google-auth-httplib2 